fix: remove useChatHistory and related things , support title
