+++ app/lib/persistence/index.ts
@@ -1,3 +1,1 @@
 export * from './localStorage';
-export * from './db';
-export * from './useChatHistory';
