--- app/lib/repoManager/types.ts
@@ -1,7 +1,0 @@
-export interface RepositoryItem {
-  projectId: string;
-  urlId: string;
-  id: string;
-  description: string;
-  timestamp: string;
-}
