+++ app/lib/stores/repo.ts
@@ -3,4 +3,5 @@ import { map } from 'nanostores';
 export const repoStore = map({
   name: '',
   path: '',
+  title: '',
 });
