+++ worker-configuration.d.ts
@@ -26,5 +26,6 @@ interface Env {
   V8_CREDIT_CLIENT_ID: string;
   V8_CREDIT_CLIENT_SECRET: string;
   VITE_ACCESS_CONTROL_ENABLED: string;
-  VITE_REPO_MANAGER_URL: string;
+  VITE_GITBASE_PERSISTENCE_ENABLED: string;
+  VITE_GITBASE_PERSISTENCE_URL: string;
 }
